+++ test/test_core.py
@@ -119,7 +119,7 @@ def test_chunk_stream():
     assert BufferedReader(chunk_stream).read() == expected
 
 
-def test_script_creator_init():
+def test_script_creator_init_with_parameters():
     script_creator = ScriptCreator(
         pathlib.Path(__file__).parent / "data" / "paramtest.ipynb"
     )
@@ -127,3 +127,9 @@ def test_script_creator_init():
         "parameter_1": (int, 6),
         "parameter_2": (str, "default value"),
     }
+
+def test_script_creator_init_no_parameters():
+    script_creator = ScriptCreator(
+        pathlib.Path(__file__).parent / "data" / "noparamtest.ipynb"
+    )
+    assert script_creator.nb_params.params == {}