+++ CHANGES.md
@@ -1,6 +1,8 @@
 ## Changes in 0.1.1 (in development)
 
 * Improve automated parameter extraction (#9)
+* Handle notebooks without parameters (#11)
+* Handle non-code cells (#12)
 
 ## Changes in 0.1.0